Add Rakefile.
Add a Rakefile to update the external components
(Zurb Foundation and Font Awesome for now). The
components can be updated individually, or you may
run

    bundle exec rake update

to update all the gems and the external components
at once.

External components are installed/downloaded
inside the lib folder. Then, files are copied to
the appropriate directories inside the source
folder.