Add the lib directory.
This directory is not strictly necessary: it is
used by 'rake update' to download/install pristine
copies of the external components, whose files are
then copied inside the source folder. Nonetheless,
it is useful to track the lib folder, so we can
check what changes in the external components.

Currently, we are committing Font Awesome 3.2.1
and the latest release of Foundation 4.