Improve support for author names.
Improve support for author names both in blog posts
and in the XML feed. Now an article's author can
be specified in the frontmatter using the 'author'
key. If no author is explicitly given, then the blog
author is used (the blog author is specified as
metadata in data/blog.yml).