Manage Font Awesome as an asset gem.
This greatly simplifies the installation and update of Font Awesome in a Middleman project. The Rake task is not needed any longer. There is no need to copy Font Awesome files in the fonts and stylesheets directories: fonts are transparently available and the line `@import "font-awesome";` in `main.css.scss` will do.

The `:require => false` bit in the Gemfile is
important (see http://middlemanapp.com/asset-pipeline/).

I have also had to remove the version constraint
for the middleman gem in the Gemfile because it was
causing errors when `bundle update` was trying to
resolve dependencies (I have no idea why).