Add support for per-page partials.
Introduce three variables: head, beginbody, and endbody, which can be used in the front matter to load additional partials on a per-page basis. For example, if you want a blog post with Mathjax, Disqus comments and Google Analytics, just add this to the front matter of the post:

    ---
    head: [mathjax]
    endbody: [disqus, analytics]
    ---

Each variable takes a list of names of partials. Such partials must exist in the /partials folder.

Of course, if a bunch of pages uses the same partials,
it is better to define a custom layout.